refactor: extract cutover/promotion into src/cutover/ (GitOps PR2 step 1)
