+++ src/cli/deploy.ts
@@ -12,17 +12,18 @@ import {
 } from "node:fs";
 import readline from "node:readline";
 import { execOrThrow, execCapture, EXEC_TIMEOUTS } from "./exec.js";
-import { readState, writeState, StateUnavailableError, type AdapterState } from "./state.js";
+import { readState, StateUnavailableError, type AdapterState } from "./state.js";
 import { discoverBuildPools, recordedBuildPools } from "./pool-topology.js";
-import {
-  cleanupRetainedStablePoolResources,
-  hasHealthCheckPolicyCrd,
-  retainRemovedPoolResources,
-} from "./stable-pool-resources.js";
-import { invalidateCdnBuildTag } from "./cdn-invalidate.js";
-import { cdnTagForBuildId } from "../cdn-tags.js";
+import { retainRemovedPoolResources } from "./stable-pool-resources.js";
+// GitOps PR2: the zero-downtime cutover (step 7's Phase D gates + Phase E promotion) lives
+// in src/cutover/. Deploy stays the orchestrating CALLER: it performs the pre-helm keep
+// transfers and the Phase-B live replica read, then hands the cutover sequence to runCutover.
+// The edge-recovery revert function is injected from ./rollback.js (below) so the module
+// boundary the orchestration tests mock stays authoritative for "revert the edge".
+import { runCutover } from "../cutover/run.js";
+import { createEdgeRecovery } from "../cutover/edge.js";
+import { CutoverExitError } from "../cutover/inputs.js";
 import { retainLiveRoutingManifest, revertRoutingServiceToBuild } from "./rollback.js";
-import { POOL_READINESS_PATH } from "../emit/templates/deployment.js";
 
 /**
  * The liveness path, used ONLY as the one-cycle fallback for the stable HealthCheckPolicy when
@@ -35,24 +36,10 @@ import { renderValkeySecret, VALKEY_SECRET_NAME } from "../emit/templates/valkey
 import { provisionMemorystore, buildDeleteMemorystoreCommand } from "./provision-cache.js";
 import { isAlreadyGoneError } from "./destroy.js";
 import { MIN_GKE_VERSION_FOR_CDN } from "./gke-version.js";
-import { routeExtJobName } from "../emit/templates/route-ext-update-job.js";
-import {
-  ROUTING_MANIFEST_SNAPSHOT_COMPONENT,
-  routingManifestSnapshotName,
-  routingServiceDeploymentName,
-} from "../emit/templates/routing-manifest-configmap.js";
-import {
-  COMPOSITION_PLAN_COMPONENT,
-  compositionPlanConfigMapName,
-} from "../emit/templates/composition-plan-configmap.js";
-// N87: internal dispatch Secrets are per BUILD and annotated `helm.sh/resource-policy: keep`,
-// so deploy owns both halves of their lifecycle — migrating the legacy stable-named one past
-// `helm upgrade`, and pruning the ones nothing references any more.
-import {
-  INTERNAL_SECRET_COMPONENT,
-  internalSecretName,
-  legacyInternalSecretName,
-} from "../emit/templates/internal-secret.js";
+// N87: internal dispatch Secrets are per BUILD and annotated `helm.sh/resource-policy: keep`.
+// Deploy owns the keep-at-upgrade half of their lifecycle (migrating the legacy stable-named
+// one past `helm upgrade`); the pruning half moved to src/cutover/gc.ts with step 7g.
+import { legacyInternalSecretName } from "../emit/templates/internal-secret.js";
 // Import the SAME sanitizer that stamps pod names and version labels (deployment.ts /
 // service.ts). The blue/green cutover patches the active Service selector to this exact
 // value, so it MUST match the pod label byte-for-byte — a divergent local copy that
@@ -71,7 +58,6 @@ import {
 } from "../emit/templates/utils.js";
 import { sanitizeForTerminal } from "./terminal.js";
 import { resolveContainerCli } from "./container-runtime.js";
-import type { TargetPlatform } from "../target-platform.js";
 import { parsePoolImageLayout } from "../pool-image-layout.js";
 // GitOps PR1: the pipeline-safe steps (A2 fingerprints, A6/A7 image build/push, A8 digest
 // resolution, B2 collision guards) live in src/pipeline/ so `emit` can run them in CI.
@@ -111,27 +97,11 @@ import {
   currentKubeContext,
   loadLocalCompositionPlan,
   preflightCompositionPlan,
-  waitForCompositionPlanReadiness,
 } from "./composition-plan.js";
 import { evaluateEnvoyGatewayPreflight } from "./envoy-gateway-preflight.js";
 
-/**
- * How long to wait for a Deployment rollout.
- *
- * 120s was arithmetically impossible against the chart's OWN rollout parameters and only ever
- * passed by luck. The emitted Deployments use `maxUnavailable: 0` with `minReadySeconds: 30` and a
- * `preStop: sleep 120` (N63, for load-balancer connection draining), so a 2-replica rollout is
- * strictly serial: each new pod takes ~10s to become ready, +30s before it counts as available,
- * and the old pod it replaces spends 120s in preStop before the next surge can start. That is
- * ~200-320s for two replicas — MEASURED on a 3-node arm64 cluster, where the routing tier
- * reliably tripped the 120s wait while `kubectl rollout status` reported success moments later.
- * GKE dodged it only because its routing Deployment spec is often unchanged between builds.
- *
- * 600s matches the chart's `progressDeadlineSeconds`: that is Kubernetes' own verdict on a stalled
- * rollout, so waiting less rejects a rollout the cluster still considers healthy, and waiting
- * longer outlives the deadline that would mark it Failed.
- */
-const KUBECTL_ROLLOUT_TIMEOUT = "--timeout=600s";
+// The 600s rollout wait (KUBECTL_ROLLOUT_TIMEOUT) and its measured-on-a-real-cluster
+// rationale moved to src/cutover/gates.ts with the Phase D gate battery (GitOps PR2).
 
 export interface DeployOptions {
   projectDir: string;
@@ -2177,75 +2147,28 @@ export async function runDeploy(options: DeployOptions): Promise<void> {
     }
   }
 
-  /**
-   * Re-point the routing tier (image + retained manifest snapshot) at the previous build after
-   * Helm was invoked. A non-zero Helm exit is not proof that nothing changed: client-side Helm
-   * can fail after applying earlier resources, and server-side apply can return an error after
-   * admission or transport uncertainty. Recovery therefore starts when mutation is attempted,
-   * not only after the command reports success.
-   */
-  let helmMutationAttempted = false;
-  const restoreEdgeToPreviousBuild = async (): Promise<{
-    attempted: boolean;
-    restored: boolean;
-    error: string;
-  }> => {
-    if (!helmMutationAttempted || !previousBuildId || previousBuildId === buildId) {
-      return { attempted: false, restored: false, error: "" };
-    }
-    try {
-      await revertRoutingServiceToBuild({
-        releaseName,
-        namespace,
-        targetBuildId: previousBuildId,
-        registry: infra.containerRegistry,
-        targetImageDigest: state?.routingImageDigests?.[previousBuildId],
-        targetPlatform: state?.targetPlatforms?.[previousBuildId],
-        // The outgoing manifest was snapshotted before Helm. Do not retain the uncertain
-        // image/manifest pair left by a failed or aborted deploy under either build's name.
-        retainCurrentManifest: false,
-      });
-      return { attempted: true, restored: true, error: "" };
-    } catch (err) {
-      return {
-        attempted: true,
-        restored: false,
-        error: err instanceof Error ? err.message : String(err),
-      };
-    }
-  };
-
-  // Human-readable tail for an abort message, describing the edge's ACTUAL state.
-  const edgeStatusLines = (r: {
-    attempted: boolean;
-    restored: boolean;
-    error: string;
-  }): string[] => {
-    if (!r.attempted) return [];
-    if (r.restored) {
-      return [
-        `  The routing edge (ext_proc: image + routing manifest) was reverted to build ` +
-          `${previousBuildId}, so edge and pools are consistent.`,
-      ];
-    }
-    return [
-      `  WARNING: could not revert the routing edge to build ${previousBuildId}: ` +
-        `${r.error || "unknown error"}`,
-      `  The edge (ext_proc) is running build ${buildId}'s middleware and routing manifest ` +
-        `while the pools serve ${previousBuildId}. Mismatched routes fall back to pool-local ` +
-        `re-resolution (invariant 1), but edge middleware is the NEW build's until repaired:`,
-      `    kubectl -n ${namespace} set image deployment/` +
-        `${routingServiceDeploymentName(releaseName)} routing-service=` +
-        `${infra.containerRegistry}/routing-service:${previousBuildId}`,
-    ];
-  };
+  // GitOps PR2: the edge-recovery pair (restoreEdgeToPreviousBuild / edgeStatusLines) moved
+  // verbatim into src/cutover/edge.ts (createEdgeRecovery). The revert function is injected
+  // from ./rollback.js so this module boundary — the one the orchestration tests mock —
+  // remains the single authority for "revert the edge".
+  const edgeRecovery = createEdgeRecovery({
+    releaseName,
+    namespace,
+    buildId,
+    previousBuildId,
+    registry: infra.containerRegistry,
+    targetImageDigest: previousBuildId ? state?.routingImageDigests?.[previousBuildId] : undefined,
+    targetPlatform: previousBuildId ? state?.targetPlatforms?.[previousBuildId] : undefined,
+    revertRoutingService: revertRoutingServiceToBuild,
+  });
+  const { restoreEdgeToPreviousBuild, edgeStatusLines } = edgeRecovery;
 
   console.log("\n  → Running helm upgrade...");
   // From this point the edge MAY run the new build. Helm overwrites the stable routing-manifest
   // ConfigMap, and a non-zero exit can still mean that write reached the API server. Every later
   // abort, including the Helm call itself, must put the edge back and report the actual result.
   if (!dryRun) {
-    helmMutationAttempted = true;
+    edgeRecovery.markHelmMutationAttempted();
     try {
       await execOrThrow("helm", helmArgs, { timeoutMs: EXEC_TIMEOUTS.cloudOperation });
     } catch (err) {
@@ -2296,1350 +2219,48 @@ export async function runDeploy(options: DeployOptions): Promise<void> {
   // pointing at a build that never took traffic, and the next deploy/rollback could delete
   // the real rollback target. State is committed only after a successful cutover (step 7d).
 
-  // 7. Zero-downtime cutover: wait for new pods, then clean up old build
+  // 7. Zero-downtime cutover: wait for new pods, then clean up old build.
+  // GitOps PR2: the whole sequence — Phase D gates (7a/7a-bis/route-ext/policy/composition/
+  // 7a-ter warm-up/7b readiness) and Phase E promotion (7c selector cutover, 7d state commit,
+  // 7d-bis warm-up restore, 7e CDN invalidation, 7f scale-down, 7g GC) — lives in
+  // src/cutover/run.ts, moved verbatim. Deploy supplies the inputs it used to read inline:
+  // the three chart-template existence probes become booleans (the cutover Job supplies them
+  // from emit-metadata instead), and previousReplicasByPool carries the Phase-B live replica
+  // read (N64). CutoverExitError maps back to the CLI exit codes the inline code used.
   if (!dryRun) {
-    // Match the new build's pods/deployments by their EXACT version label — the same value the
-    // cutover patches Services to (`sanitizeK8sName(buildId)`, which stamps `app.kubernetes.io/
-    // version` on every pod). The prior match used a 12-char normalized-prefix substring, so an OLD
-    // build whose id shared that prefix could satisfy this readiness check; the cutover would then
-    // patch Services to the full new label, match zero pods, drain the NEG, and 503 the origin.
-    const safeBuildId = sanitizeK8sName(buildId);
-
-    // 7a. Wait for new deployment to be ready
-    console.log(`\n  → Waiting for new pods to be ready...`);
-    const newDeployResult = await execCapture(
-      "kubectl",
-      [
-        "get",
-        "deployments",
-        "-n",
-        namespace,
-        "-l",
-        `app.kubernetes.io/name=${releaseName},app.kubernetes.io/version=${safeBuildId}`,
-        "-o",
-        'jsonpath={range .items[*]}{.metadata.name}{"\\n"}{end}',
-      ],
-      { timeoutMs: EXEC_TIMEOUTS.kubectl },
-    );
-    // The routing service carries the same version label but is a stable in-place Deployment,
-    // verified separately below — exclude it here by EXACT name (a substring match would
-    // also exclude a pool deployment that merely contains "routing-service" in its name).
-    const routingDeploy = routingServiceDeploymentName(releaseName);
-    const newDeploys = (newDeployResult.stdout?.trim().split("\n") ?? []).filter(
-      (n) => n && n !== routingDeploy,
-    );
-
-    for (const deployName of newDeploys) {
-      console.log(`    Waiting for ${deployName}...`);
-      const rollout = await execCapture(
-        "kubectl",
-        ["rollout", "status", `deployment/${deployName}`, "-n", namespace, KUBECTL_ROLLOUT_TIMEOUT],
-        { timeoutMs: EXEC_TIMEOUTS.rollout },
-      );
-      // The pod-health gate below is the real readiness gate, but don't walk into it on
-      // an already-failed rollout — the exit code was previously discarded, so a stuck
-      // deployment burned the whole 2-minute health budget before failing.
-      if (rollout.exitCode !== 0) {
-        // N25: the edge already runs the new build (helm overwrote the stable manifest
-        // ConfigMap) — put it back before claiming the previous build is still serving.
-        const edge = await restoreEdgeToPreviousBuild();
-        throw new Error(
-          [
-            `Deployment ${deployName} did not finish rolling out within 120s. Traffic was ` +
-              `NOT switched — the previous build's pools are still serving.`,
-            // L14: kubectl rollout output carries controller/admission messages.
-            `${sanitizeForTerminal((rollout.stderr || rollout.stdout).trim())}`,
-            `Inspect: kubectl logs deployment/${deployName} -n ${namespace} --tail=40`,
-            ...edgeStatusLines(edge),
-          ].join("\n"),
-        );
-      }
-    }
-
-    // 7a-bis. The routing service (ext_proc edge) is a stable Deployment updated in place
-    // per build, and was historically excluded from readiness. That let a broken image
-    // (e.g. a missing @next/routing) crashloop undetected for months while Kubernetes kept
-    // the old ReplicaSet serving stale edge code and the deploy reported success. Verify it
-    // actually rolls out; a stuck rollout is fatal.
-    const rsExists = await execCapture(
-      "kubectl",
-      ["get", "deployment", routingDeploy, "-n", namespace, "--ignore-not-found", "-o", "name"],
-      { timeoutMs: EXEC_TIMEOUTS.kubectl },
-    );
-    if (rsExists.exitCode === 0 && rsExists.stdout.trim()) {
-      console.log(`    Waiting for ${routingDeploy}...`);
-      const rsRollout = await execCapture(
-        "kubectl",
-        [
-          "rollout",
-          "status",
-          `deployment/${routingDeploy}`,
-          "-n",
-          namespace,
-          KUBECTL_ROLLOUT_TIMEOUT,
-        ],
-        { timeoutMs: EXEC_TIMEOUTS.rollout },
-      );
-      if (rsRollout.exitCode !== 0) {
-        // The new routing pods can't roll — but the OLD routing pods already picked up the
-        // new build's manifest from the (helm-overwritten) stable ConfigMap via kubelet
-        // volume sync, so the edge is on the new build regardless. Revert it (N25).
-        const edge = await restoreEdgeToPreviousBuild();
-        throw new Error(
-          [
-            `Routing service (${routingDeploy}) did not become healthy. Traffic was NOT ` +
-              `switched — the previous build's pools are still serving.`,
-            `${(rsRollout.stderr || rsRollout.stdout).trim()}`,
-            `Inspect: kubectl logs -l app.kubernetes.io/component=routing-service -n ${namespace} --tail=40`,
-            ...edgeStatusLines(edge),
-          ].join("\n"),
-        );
-      }
-    }
-
-    // The traffic extension is part of the middleware security boundary. Reconcile and verify
-    // it while the active Services still select the previous build; never cut traffic and call
-    // the deploy successful with a missing/incomplete ext_proc backend.
-    const currentRouteExtJob = routeExtJobName(releaseName, buildId);
-    const routeExtJobYaml = path.join(outputDir, "chart", "templates", "route-ext-update-job.yaml");
-    if (existsSync(routeExtJobYaml)) {
-      const wait = await execCapture(
-        "kubectl",
-        [
-          "wait",
-          "--for=condition=complete",
-          `job/${currentRouteExtJob}`,
-          "-n",
-          namespace,
-          "--timeout=600s",
-        ],
-        { timeoutMs: EXEC_TIMEOUTS.rollout },
-      );
-      if (wait.exitCode !== 0) {
-        const edge = await restoreEdgeToPreviousBuild();
-        throw new Error(
-          [
-            `ext_proc registration job (${currentRouteExtJob}) did not complete; refusing ` +
-              `traffic cutover because middleware may not be wired.`,
-            // L14: kubectl wait output carries controller/admission messages.
-            `${sanitizeForTerminal((wait.stderr || wait.stdout).trim())}`,
-            `Inspect: kubectl logs job/${currentRouteExtJob} -n ${namespace}`,
-            ...edgeStatusLines(edge),
-          ].join("\n"),
-        );
-      }
-      console.log("  → ext_proc traffic extension registration job completed ✓");
-    }
-
-    // The SAME boundary, for providers that register ext_proc in-cluster. There is no Job to
-    // wait on, so before this the generic path verified NOTHING: an EnvoyExtensionPolicy with
-    // Accepted=False, a GatewayClass whose controller is not Envoy, or label drift on the proxy
-    // selector would all leave traffic 500ing (fail-closed) or silently routed without the edge
-    // tier — while the deploy printed success. Middleware not running is the exact class of
-    // failure this gate exists for, so it must cover both registration mechanisms.
-    const policyYaml = path.join(outputDir, "chart", "templates", "envoy-extension-policy.yaml");
-    if (existsSync(policyYaml)) {
-      const policyName = `${releaseName}-routing-extproc`;
-      // POLL, and require the condition to be CURRENT for the object's generation.
-      //
-      // A single immediate read was wrong twice over: a freshly reconciled policy may have no
-      // status yet (false abort), and an UPDATED policy retains the previous
-      // `Accepted=True` until the controller catches up (false pass — the dangerous direction,
-      // since it green-lights a cutover whose edge may not be wired). `observedGeneration`
-      // vs `metadata.generation` is what distinguishes "accepted" from "accepted, previously".
-      //
-      // Ancestors are a map-like list, not an ordered one, so `[0]` is not necessarily this
-      // release's Gateway/route — select by the Envoy Gateway controller instead.
-      let status = "";
-      let detail = "";
-      for (let attempt = 0; attempt < 30; attempt++) {
-        const read = await execCapture(
-          "kubectl",
-          ["get", "envoyextensionpolicy", policyName, "-n", namespace, "-o", "json"],
-          { timeoutMs: EXEC_TIMEOUTS.kubectl },
-        );
-        detail = (read.stderr || "").trim();
-        if (read.exitCode === 0) {
-          try {
-            const obj = JSON.parse(read.stdout) as {
-              metadata?: { generation?: number };
-              status?: {
-                ancestors?: Array<{
-                  controllerName?: string;
-                  conditions?: Array<{
-                    type?: string;
-                    status?: string;
-                    observedGeneration?: number;
-                  }>;
-                }>;
-              };
-            };
-            const generation = obj.metadata?.generation;
-            const ancestors = obj.status?.ancestors ?? [];
-            const mine =
-              ancestors.find((a) => a.controllerName?.includes("gateway.envoyproxy.io")) ??
-              ancestors[0];
-            const cond = mine?.conditions?.find((c) => c.type === "Accepted");
-            const current =
-              cond !== undefined &&
-              (generation === undefined ||
-                cond.observedGeneration === undefined ||
-                cond.observedGeneration >= generation);
-            if (cond && current) {
-              status = cond.status ?? "";
-              if (status === "True") break;
-              // A definitive False is final — waiting will not fix a rejected policy.
-              if (status === "False") break;
-            }
-          } catch {
-            // Unparseable status: keep polling rather than treating it as a verdict.
-          }
-        }
-        await new Promise((r) => setTimeout(r, 2000));
-      }
-      if (status !== "True") {
-        const edge = await restoreEdgeToPreviousBuild();
-        throw new Error(
-          [
-            `ext_proc EnvoyExtensionPolicy (${policyName}) is not Accepted (status ` +
-              `${status || "unknown"}); refusing traffic cutover because middleware would not ` +
-              `run at the edge.`,
-            sanitizeForTerminal(detail),
-            `Inspect: kubectl describe envoyextensionpolicy ${policyName} -n ${namespace}`,
-            `Common causes: the GatewayClass controller is not Envoy Gateway (an ` +
-              `EnvoyExtensionPolicy only applies to one), or the Gateway it targets does not exist.`,
-            ...edgeStatusLines(edge),
-          ].join("\n"),
-        );
-      }
-      console.log("  → ext_proc EnvoyExtensionPolicy accepted ✓");
-    }
-
-    if (compositionSnapshot) {
-      console.log("  → Verifying composition-plan readiness...");
-      try {
-        await waitForCompositionPlanReadiness(compositionSnapshot.plan);
-      } catch (error) {
-        const edge = await restoreEdgeToPreviousBuild();
-        throw new Error(
-          [
-            `Composition-plan readiness failed; refusing traffic cutover: ` +
-              `${error instanceof Error ? error.message : String(error)}`,
-            ...edgeStatusLines(edge),
-          ].join("\n"),
-        );
-      }
-      console.log("  → Composition-plan resources are ready ✓");
-    }
-
-    // 7a-ter. N64: match the outgoing build's CAPACITY before the gate below can pass.
-    // The chart renders a new build at its HPA floor (`replicas.min`), while the build
-    // being replaced may be sitting far above that after autoscaling under load. Cutting
-    // over then hands 100% of traffic to `min` pods with the HPA climbing from behind —
-    // and the gate cannot simply WAIT for the higher count, because with no traffic yet
-    // nothing would ever scale the new build up. So ask for the capacity explicitly here;
-    // the HPA takes over from real load once traffic arrives, and the next helm upgrade
-    // re-renders the floor.
-    //
-    // N67: `helm upgrade` above ALSO installed the new build's own HPA
-    // (`<release>-<pool>-<build>-hpa`) with the chart's `replicas.min`/`replicas.max`. When
-    // the outgoing build sits ABOVE the new chart's `max` — it autoscaled under load, or a
-    // rollback widened it (N26), or the operator lowered the ceiling — the autoscaler
-    // reconciles the manual scale below straight back down within one control loop, and the
-    // gate at 7b then waits for a count that can NEVER be reached: the deploy burned the
-    // whole health budget and aborted BEFORE cutover, i.e. the capacity gate blocked every
-    // such deploy. Raising only maxReplicas is not enough: an idle new build has no load,
-    // so the HPA's desired count remains its configured minReplicas and it can reconcile a
-    // manual `kubectl scale` straight back to that floor. Do not assume a scale-down
-    // stabilization window protects the warm-up — that behavior is configurable and is not
-    // part of the generated HPA. Temporarily lift BOTH bounds around the capacity gate,
-    // remember their exact chart-rendered values, and put both back on EVERY exit path
-    // (success AND abort). After cutover, real load under the chart's own bounds decides the
-    // count again.
-    const warmedHpas: { name: string; min: number; max: number }[] = [];
-    const restoredWarmedHpas = new Set<string>();
-    const restoreWarmedHpas = async (): Promise<void> => {
-      for (const { name, min, max } of warmedHpas) {
-        if (restoredWarmedHpas.has(name)) continue;
-        let lastFailure = "";
-        for (let attempt = 1; attempt <= 3; attempt++) {
-          const restore = await execCapture(
-            "kubectl",
-            [
-              "patch",
-              "hpa",
-              name,
-              "-n",
-              namespace,
-              "--type=merge",
-              // Same field-manager rationale as the Service/Deployment patches: helm owns the
-              // chart-rendered HPA, so the next `helm upgrade` must not conflict here.
-              "--field-manager=helm",
-              "-p",
-              JSON.stringify({ spec: { minReplicas: min, maxReplicas: max } }),
-            ],
-            { timeoutMs: EXEC_TIMEOUTS.kubectl },
-          );
-          if (restore.exitCode === 0) {
-            restoredWarmedHpas.add(name);
-            console.log(
-              `  → Restored ${name} to the chart's minReplicas=${min}, maxReplicas=${max}`,
-            );
-            break;
-          }
-          lastFailure = sanitizeForTerminal(restore.stderr.trim()) || `exit ${restore.exitCode}`;
-        }
-        if (!restoredWarmedHpas.has(name)) {
-          console.warn(
-            `  ! Could not restore ${name} to the chart's minReplicas=${min}, ` +
-              `maxReplicas=${max} after 3 attempts (${lastFailure}) — ` +
-              `it still has the temporary pre-cutover bounds, so this pool may retain warm-up ` +
-              `capacity until the next \`helm upgrade\` re-renders it. Fix it with: kubectl -n ` +
-              `${namespace} patch hpa ${name} --type=merge -p ` +
-              `'{"spec":{"minReplicas":${min},"maxReplicas":${max}}}'`,
-          );
-        }
-      }
-    };
-
-    // A partial warm-up setup cannot safely continue. If the HPA cannot be read, its
-    // original bounds cannot be restored; if the temporary patch is rejected, it may
-    // immediately lower the idle Deployment again. Restore every pool already prepared and
-    // put ext_proc back on the serving build before failing, without waiting out the health
-    // budget for a capacity target that is not stable.
-    const abortWarmupSetup = async (message: string): Promise<never> => {
-      const edge = await restoreEdgeToPreviousBuild();
-      await restoreWarmedHpas();
-      throw new Error([message, ...edgeStatusLines(edge)].join("\n"));
-    };
-
-    // S18: seeded from EVERY configured pool at a floor of one, not only from pools with a
-    // live predecessor. previousReplicasByPool has no entry for a pool that is new in this
-    // build, nor for any pool on a first deploy, so those pools used to contribute no
-    // expectation at all: a sibling's ready pods satisfied `checkedCount > 0`, the gate
-    // passed, and cutover then patched EVERY active Service to the new build — leaving the
-    // new pool's Service with zero endpoints and serving 503s. One ready pod is the weakest
-    // claim worth making ("something in this pool is actually serving"), and for pools that
-    // DO have a predecessor the real count below overwrites it immediately.
-    const capacityTargets = new Map<string, number>(pools.map((poolName) => [poolName, 1]));
-    for (const [poolName, replicas] of previousReplicasByPool) {
-      // A removed/renamed pool has no incoming Deployment to warm. It remains in
-      // previousReplicasByPool because deploy must preserve and later park it, but adding it
-      // here creates an impossible health target and makes every topology-changing deploy time
-      // out before cutover. Only common pools inherit outgoing capacity; newly-added pools keep
-      // the one-ready-pod floor above.
-      if (capacityTargets.has(poolName)) capacityTargets.set(poolName, replicas);
-    }
-
-    const shortfalls: string[] = [];
-    for (const poolName of pools) {
-      const outgoing = previousReplicasByPool.get(poolName);
-      const { hpa: newHpaName } = poolResourceNames(releaseName, poolName, buildId);
-
-      // The chart always renders one HPA per pool. A missing HPA after Helm means the release
-      // was only partially applied; accepting it would cut traffic to an unautoscaled pool.
-      // Probe every pool, including first-deploy and newly added pools, before scaling any of
-      // them. `--ignore-not-found` keeps absence machine-readable without hiding read errors.
-      const hpaRead = await execCapture(
-        "kubectl",
-        [
-          "get",
-          "hpa",
-          newHpaName,
-          "-n",
-          namespace,
-          "--ignore-not-found",
-          "-o",
-          "jsonpath={.metadata.name}|{.spec.minReplicas}|{.spec.maxReplicas}",
-        ],
-        { timeoutMs: EXEC_TIMEOUTS.kubectl },
-      );
-      if (hpaRead.exitCode !== 0) {
-        await abortWarmupSetup(
-          `Could not read the new build's HPA ${newHpaName} (kubectl exited ` +
-            `${hpaRead.exitCode}${hpaRead.stderr.trim() ? `: ${sanitizeForTerminal(hpaRead.stderr.trim())}` : ""}). ` +
-            `Refusing to warm pool "${poolName}" because its original bounds cannot be ` +
-            `preserved or its capacity kept stable. Nothing was cut over.`,
-        );
-      }
-
-      const [hpaFound, hpaMinField, hpaMaxField] = hpaRead.stdout.trim().split("|");
-      if (!hpaFound) {
-        await abortWarmupSetup(
-          `The new build's expected HPA ${newHpaName} was not found after Helm applied the ` +
-            `release. Refusing to cut over pool "${poolName}" to a partially applied build. ` +
-            `Nothing was cut over.`,
-        );
-      }
-      const hpaMin = Number(hpaMinField);
-      const hpaMax = Number(hpaMaxField);
-      if (!Number.isInteger(hpaMin) || hpaMin < 1 || !Number.isInteger(hpaMax) || hpaMax < hpaMin) {
-        await abortWarmupSetup(
-          `Could not parse the new build's HPA ${newHpaName} bounds ` +
-            `(minReplicas=${JSON.stringify(hpaMinField ?? "")}, ` +
-            `maxReplicas=${JSON.stringify(hpaMaxField ?? "")}). Refusing to warm pool ` +
-            `"${poolName}" because its original bounds cannot be restored safely. ` +
-            `Nothing was cut over.`,
-        );
-      }
-
-      if (outgoing !== undefined && hpaMin < outgoing) {
-        const warmMin = outgoing;
-        const warmMax = Math.max(hpaMax, warmMin);
-        console.log(
-          `  → Warming ${newHpaName} at minReplicas=${warmMin}, ` +
-            `maxReplicas=${warmMax} so the idle new pool keeps the outgoing build's ` +
-            `${outgoing} replicas (restored to minReplicas=${hpaMin}, ` +
-            `maxReplicas=${hpaMax} after cutover)`,
-        );
-        const warm = await execCapture(
-          "kubectl",
-          [
-            "patch",
-            "hpa",
-            newHpaName,
-            "-n",
-            namespace,
-            "--type=merge",
-            "--field-manager=helm",
-            "-p",
-            JSON.stringify({ spec: { minReplicas: warmMin, maxReplicas: warmMax } }),
-          ],
-          { timeoutMs: EXEC_TIMEOUTS.kubectl },
-        );
-        if (warm.exitCode !== 0) {
-          await abortWarmupSetup(
-            `Could not set temporary warm-up bounds on ${newHpaName} ` +
-              `(minReplicas=${warmMin}, maxReplicas=${warmMax}; ` +
-              `${sanitizeForTerminal(warm.stderr.trim()) || `kubectl exited ${warm.exitCode}`}). Refusing to ` +
-              `scale or cut over pool "${poolName}" because its HPA could immediately ` +
-              `remove the required capacity. Nothing was cut over.`,
-          );
-        }
-        warmedHpas.push({ name: newHpaName, min: hpaMin, max: hpaMax });
-      }
-    }
-
-    // Prepare every autoscaler before scaling any Deployment. A read or admission failure
-    // on a later pool then restores the earlier HPA patches without leaving part of the new
-    // build manually scaled above its chart state.
-    for (const [poolName, outgoing] of previousReplicasByPool) {
-      const { deployment: newDeployName } = poolResourceNames(releaseName, poolName, buildId);
-      const expected = capacityTargets.get(poolName) ?? outgoing;
-      const cur = await execCapture(
-        "kubectl",
-        [
-          "get",
-          "deployment",
-          newDeployName,
-          "-n",
-          namespace,
-          "--ignore-not-found",
-          "-o",
-          "jsonpath={.metadata.name}|{.spec.replicas}",
-        ],
-        { timeoutMs: EXEC_TIMEOUTS.kubectl },
-      );
-      const [foundName, replicasField] = cur.stdout.trim().split("|");
-      const current = parseInt(replicasField ?? "", 10);
-      if (cur.exitCode !== 0 || !foundName || !Number.isFinite(current)) {
-        // Don't abort here — the readiness gate below is the authority and will fail
-        // closed if the capacity never materializes. Say why we could not pre-scale.
-        console.warn(
-          `  ! Could not read the new build's replica count for ${newDeployName} ` +
-            `(kubectl exited ${cur.exitCode}${cur.stderr.trim() ? `: ${sanitizeForTerminal(cur.stderr.trim())}` : ""}) ` +
-            `— skipping the pre-cutover scale-up to the outgoing build's ${expected} replicas.`,
-        );
-        continue;
-      }
-      if (current >= expected) continue;
-      console.log(
-        `  → Matching outgoing capacity for pool "${poolName}": scaling ${newDeployName} ` +
-          `${current} → ${expected} replicas before cutover`,
-      );
-      const scaleUp = await execCapture(
-        "kubectl",
-        ["scale", `deployment/${newDeployName}`, "-n", namespace, `--replicas=${expected}`],
-        { timeoutMs: EXEC_TIMEOUTS.kubectl },
-      );
-      if (scaleUp.exitCode !== 0) {
-        shortfalls.push(
-          `${newDeployName}: could not scale to ${expected} ` +
-            `(${sanitizeForTerminal(scaleUp.stderr.trim()) || `exit ${scaleUp.exitCode}`})`,
-        );
-      }
-    }
-    if (shortfalls.length > 0) {
-      // Not fatal on its own: the gate below requires readyReplicas >= the outgoing count
-      // per pool, so an unfulfilled scale-up aborts the cutover there with full context.
-      for (const s of shortfalls) console.warn(`  ! ${s}`);
-    }
-
-    // 7b. Wait for the new build's pods to be READY from inside the cluster.
-    // This reads each pod's `Ready` condition — which the kubelet drives from the
-    // readinessProbe, and that probe now targets `/readyz` (N32), a 503-until-serving
-    // endpoint (instrumentation registered, at least one route module loaded) rather than
-    // the hardcoded-200 `/healthz` that could not fail. GCP LB backend health is
-    // deliberately NOT waited on here: backend propagation can take 5+ minutes for a new
-    // backend, and the NEG health checks gate the endpoints independently after the flip.
-    //
-    // N64: readiness of SOME pods is not enough — the gate requires at least as many ready
-    // pods per pool as the build being replaced is running (previousReplicasByPool). It
-    // used to pass on `checkedCount > 0`, i.e. ONE ready pod, so a previous build serving 6
-    // cut over to a single pod.
-    console.log(`  → Verifying new pods are serving...`);
-    let newBuildHealthy = false;
-    let lastShortfall = "";
-    const maxHealthAttempts = 24; // 2 minutes (5s intervals)
-    for (let attempt = 0; attempt < maxHealthAttempts; attempt++) {
-      let allHealthy = true;
-      let checkedCount = 0;
-      const readyByPool = new Map<string, number>();
-      const podsResult = await execCapture(
-        "kubectl",
-        [
-          "get",
-          "pods",
-          "-n",
-          namespace,
-          "-l",
-          // The routing service shares the version label; exclude it by its component label
-          // (exact — a name substring match would also drop pool pods named similarly).
-          `app.kubernetes.io/name=${releaseName},app.kubernetes.io/version=${safeBuildId},app.kubernetes.io/component!=routing-service`,
-          "-o",
-          // The component label IS the pool name (deployment.ts stamps it), which is what
-          // makes the per-pool count below exact without parsing pod names.
-          'jsonpath={range .items[*]}{.metadata.name}|{.status.conditions[?(@.type=="Ready")].status}|{.metadata.labels.app\\.kubernetes\\.io/component}{"\\n"}{end}',
-        ],
-        { timeoutMs: EXEC_TIMEOUTS.kubectl },
-      );
-      if (podsResult.exitCode === 0) {
-        for (const line of podsResult.stdout.trim().split("\n")) {
-          const [podName, ready, component] = line.split("|");
-          if (!podName) continue; // selector already scoped to this build's pool pods
-          checkedCount++;
-          if (ready !== "True") {
-            allHealthy = false;
-            continue;
-          }
-          if (component) readyByPool.set(component, (readyByPool.get(component) ?? 0) + 1);
-        }
-      }
-      // Per-pool capacity check against the build being replaced.
-      const missing: string[] = [];
-      for (const [poolName, expected] of capacityTargets) {
-        const ready = readyByPool.get(poolName) ?? 0;
-        if (ready < expected) missing.push(`${poolName}: ${ready}/${expected} ready`);
-      }
-      lastShortfall = missing.join(", ");
-      if (allHealthy && checkedCount > 0 && missing.length === 0) {
-        console.log(`    All ${checkedCount} new pods ready and serving`);
-        newBuildHealthy = true;
-        break;
-      }
-      if (attempt < maxHealthAttempts - 1) {
-        await new Promise((r) => setTimeout(r, 5000));
-      }
-    }
-
-    if (!newBuildHealthy) {
-      // N25 before anything else: the ext_proc edge has been on the new build since
-      // `helm upgrade` (stable manifest ConfigMap + in-place routing Deployment), so
-      // "no cutover performed" was only ever true of the POOLS.
-      const edge = await restoreEdgeToPreviousBuild();
-      // N67: and both temporary warm-up bounds go back before we leave — this build is
-      // being abandoned, so it must not keep the raised replica floor.
-      await restoreWarmedHpas();
-      console.error(`\n  DEPLOY FAILED: New build did not become healthy within 2 minutes.`);
-      console.error(
-        `  The previous build's pools are still serving traffic. No pool cutover was performed.`,
-      );
-      for (const line of edgeStatusLines(edge)) console.error(line);
-      if (lastShortfall) {
-        // N64: distinguish "pods aren't ready" from "not ENOUGH pods are ready" — the
-        // latter is a capacity gate, and the numbers say which pool fell short of the
-        // outgoing build's live count.
-        console.error(
-          `  Capacity gate: the new build must match the outgoing build's live replica ` +
-            `count per pool before cutover — ${lastShortfall}.`,
-        );
-      }
-      console.error("");
-
-      // Try to get more diagnostic info
-      const newPods = await execCapture(
-        "kubectl",
-        [
-          "get",
-          "pods",
-          "-n",
-          namespace,
-          "-l",
-          `app.kubernetes.io/name=${releaseName},app.kubernetes.io/version=${safeBuildId},app.kubernetes.io/component!=routing-service`,
-          "-o",
-          'jsonpath={range .items[*]}{.metadata.name}|{.status.phase}{"\\n"}{end}',
-        ],
-        { timeoutMs: EXEC_TIMEOUTS.kubectl },
-      );
-      if (newPods.exitCode === 0 && newPods.stdout.trim()) {
-        const podLines = newPods.stdout.trim().split("\n");
-        for (const line of podLines) {
-          const [podName, phase] = line.split("|");
-          if (!podName) continue; // selector already scoped to this build's pool pods
-          console.error(`  Pod ${podName}: ${phase}`);
-          // N32: probe the READINESS endpoint, not `/healthz`. `/healthz` returns a
-          // hardcoded 200 before any routing or handler load — it cannot fail, so asking it
-          // here told the operator nothing at exactly the moment they need something.
-          // `/readyz` 503s until the pod can actually serve and its body carries a `reason`
-          // ("instrumentation register() failed", "route module /x failed to load"), which
-          // is the answer to "why is this pod not Ready?" — so print the body too.
-          const readyzResult = await execCapture(
-            "kubectl",
-            [
-              "exec",
-              podName,
-              "-n",
-              namespace,
-              "--",
-              "node",
-              "-e",
-              `const http=require("http");http.get("http://localhost:3000${POOL_READINESS_PATH}",r=>{let d="";r.on("data",c=>d+=c);r.on("end",()=>console.log(r.statusCode,d))}).on("error",e=>console.log("ERR",e.message))`,
-            ],
-            { timeoutMs: EXEC_TIMEOUTS.kubectl },
-          );
-          if (readyzResult.exitCode === 0 && readyzResult.stdout.trim()) {
-            // L14: the body is pod-sourced text — strip terminal control characters.
-            console.error(
-              `  Readiness (${POOL_READINESS_PATH}): ` +
-                `${sanitizeForTerminal(readyzResult.stdout.trim()).slice(0, 300)}`,
-            );
-          } else if (readyzResult.exitCode !== 0) {
-            console.error(
-              `  Readiness (${POOL_READINESS_PATH}): probe could not run ` +
-                `(kubectl exec exited ${readyzResult.exitCode}) — the container may not be ` +
-                `running yet; see the pod logs below.`,
-            );
-          }
-          // Get last error from logs
-          const logsResult = await execCapture(
-            "kubectl",
-            ["logs", podName, "-n", namespace, "--tail=20"],
-            { timeoutMs: EXEC_TIMEOUTS.kubectl },
-          );
-          if (logsResult.exitCode === 0) {
-            const errorLines = logsResult.stdout
-              .split("\n")
-              .filter(
-                (l) =>
-                  l.includes("Error") ||
-                  l.includes("error") ||
-                  l.includes("FATAL") ||
-                  l.includes("Cannot find"),
-              );
-            if (errorLines.length > 0) {
-              console.error(`  Errors:`);
-              for (const err of errorLines.slice(0, 5)) {
-                // L14: pod log lines are cluster-sourced — strip terminal control
-                // characters before printing.
-                console.error(`    ${sanitizeForTerminal(err.trim()).slice(0, 150)}`);
-              }
-            } else {
-              console.error(
-                `  No errors in pod logs. The issue may be GCP health check configuration.`,
-              );
-            }
-          }
-        }
-      }
-
-      console.error(`\n  Diagnose:  npx adapter-k8s doctor`);
-      console.error(`  Tail logs: npx adapter-k8s tail`);
-      process.exit(1);
-    }
-
-    // 7c. Cut traffic over: patch each active Service selector to the new build.
-    // MUST use the exact same sanitizer that stamped the pod labels (sanitizeK8sName,
-    // which prepends `b-` when the build id starts with a non-letter). An inline
-    // transform that omits the `b-` prefix writes a selector that matches no pods:
-    // the Service drains to zero endpoints, its standalone NEG empties, and the LB
-    // returns 503 `failed_to_connect_to_backend` for every origin request (only CDN
-    // cache hits survive). This bit us on build ids beginning with a digit.
-    // (safeBuildId is computed once in step 7 above and reused here.)
-    console.log(`  → Switching traffic to new build...`);
-    const patchFailures: { pool: string; service: string; stderr: string }[] = [];
-    const patchedServices: string[] = [];
-    const originalServiceSelectors = new Map<string, Record<string, string>>();
-
-    // A topology-changing rollback may have redirected a stable Service to a fallback pool.
-    // Read every selector before changing any of them so cutover restores both component and
-    // version, and a partial failure can put the exact live selectors back.
-    const serviceDestinations = [
-      ...pools.map((pool) => ({ servicePool: pool, targetPool: pool })),
-      ...(hasPortableOrigin ? [{ servicePool: "origin", targetPool: defaultPool }] : []),
-    ];
-    for (const { servicePool } of serviceDestinations) {
-      const activeServiceName = sanitizeK8sName(`${releaseName}-${servicePool}`);
-      const read = await execCapture(
-        "kubectl",
-        ["get", "service", activeServiceName, "-n", namespace, "-o", "json"],
-        { timeoutMs: EXEC_TIMEOUTS.kubectl },
-      );
-      let selector: unknown;
-      if (read.exitCode === 0) {
-        try {
-          selector = JSON.parse(read.stdout)?.spec?.selector;
-        } catch {
-          selector = undefined;
-        }
-      }
-      if (
-        !selector ||
-        typeof selector !== "object" ||
-        Array.isArray(selector) ||
-        Object.values(selector).some((value) => typeof value !== "string")
-      ) {
-        patchFailures.push({
-          pool: servicePool,
-          service: activeServiceName,
-          stderr:
-            read.stderr.trim() ||
-            `could not read an exact Service selector (kubectl exited ${read.exitCode})`,
-        });
-        continue;
-      }
-      originalServiceSelectors.set(activeServiceName, selector as Record<string, string>);
-    }
-
-    if (patchFailures.length === 0) {
-      for (const { servicePool, targetPool } of serviceDestinations) {
-        const activeServiceName = sanitizeK8sName(`${releaseName}-${servicePool}`);
-        const originalSelector = originalServiceSelectors.get(activeServiceName)!;
-        const patchResult = await execCapture(
-          "kubectl",
-          [
-            "patch",
-            "service",
-            activeServiceName,
-            "-n",
-            namespace,
-            "--type=json",
-            // Keep this imperative cutover under helm's field manager. On Helm 4's server-side
-            // path that prevents the next upgrade conflicting on the selector we flip here;
-            // Helm 3's client-side merge does not enforce managed-field conflicts.
-            // NOTE: --force-conflicts is NOT a valid `kubectl patch` flag (it only exists on
-            // `kubectl apply --server-side`); a JSON patch is not server-side apply and needs
-            // no conflict override.
-            "--field-manager=helm",
-            "-p",
-            JSON.stringify([
-              {
-                op: "replace",
-                path: "/spec/selector",
-                value: {
-                  ...originalSelector,
-                  "app.kubernetes.io/component": targetPool,
-                  "app.kubernetes.io/version": safeBuildId,
-                },
-              },
-            ]),
-          ],
-          { timeoutMs: EXEC_TIMEOUTS.kubectl },
-        );
-        if (patchResult.exitCode !== 0) {
-          patchFailures.push({
-            pool: servicePool,
-            service: activeServiceName,
-            stderr: patchResult.stderr.trim(),
-          });
-        } else {
-          patchedServices.push(activeServiceName);
-        }
-      }
-    }
-
-    // If ANY pool's selector patch failed, some/all Services still point at the old
-    // build. Deleting old deployments now would strand those Services with zero healthy
-    // endpoints. Abort the cutover, leave the previous build in place, and fail loudly
-    // rather than proceeding to the cleanup below and printing "Deploy complete".
-    if (patchFailures.length > 0) {
-      const revertFailures: string[] = [];
-      for (const serviceName of patchedServices) {
-        const originalSelector = originalServiceSelectors.get(serviceName)!;
-        const revertResult = await execCapture(
-          "kubectl",
-          [
-            "patch",
-            "service",
-            serviceName,
-            "-n",
-            namespace,
-            "--type=json",
-            "--field-manager=helm",
-            "-p",
-            JSON.stringify([
-              {
-                op: "replace",
-                path: "/spec/selector",
-                value: originalSelector,
-              },
-            ]),
-          ],
-          { timeoutMs: EXEC_TIMEOUTS.kubectl },
-        );
-        if (revertResult.exitCode !== 0) revertFailures.push(serviceName);
-      }
-      // N25: the pools stay on the previous build, so the edge must too.
-      const edge = await restoreEdgeToPreviousBuild();
-      // N67: restore the chart's autoscaling bounds on the build we are abandoning.
-      await restoreWarmedHpas();
-      console.error(`\n  DEPLOY FAILED: traffic was NOT switched to the new build.`);
-      console.error(
-        `  ${patchFailures.length} of ${serviceDestinations.length} Service selector patch(es) failed:`,
-      );
-      for (const f of patchFailures) {
-        console.error(
-          // L14: the patch stderr is apiserver-sourced.
-          `    - pool "${f.pool}" (service ${f.service}): ` +
-            `${sanitizeForTerminal(f.stderr) || "unknown error"}`,
-        );
-      }
-      if (revertFailures.length > 0) {
-        console.error(
-          `  WARNING: failed to restore selector(s) for: ${revertFailures.join(", ")}.`,
-        );
-        console.error(`  Traffic may be split across builds; repair those Services manually.`);
-      } else if (patchedServices.length > 0) {
-        console.error(`  Any successful selector patches were restored to their prior values.`);
-      }
-      console.error(`  Old deployments were left in place.`);
-      for (const line of edgeStatusLines(edge)) console.error(line);
-      console.error(`  No cleanup was performed. Investigate and re-run the deploy.\n`);
-      console.error(`  Diagnose:  npx adapter-k8s doctor`);
-      // Do NOT write committed state here: traffic did not switch, so the previously-serving
-      // build is still current. Recording the new build as current would strand the real
-      // rollback target on the next deploy/rollback.
-      process.exit(1);
-    }
-
-    // 7d. Commit deploy state IMMEDIATELY after the confirmed cutover — before anything
-    // else post-switch. Every extra step between the selector patch and this write is a
-    // Ctrl-C window in which traffic serves the new build while state still names the old
-    // one (the next deploy/rollback would then target the wrong build). Committing ONLY
-    // after a confirmed cutover still holds: a failed deploy never records a never-serving
-    // build as current. Traffic has already switched; if the cluster ConfigMap write fails
-    // we surface it loudly (local was updated) rather than silently diverging.
     try {
-      // M13: record the exact Cache-Tag THIS build's pool-server stamps (derived HERE,
-      // at its deploy — the only moment code and pods agree), and carry the outgoing
-      // build's recorded tag verbatim (it stays the rollback target; re-deriving its tag
-      // under newer code is exactly the M13 failure). Pruned to the two builds in play.
-      const cdnTags: Record<string, string> = {};
-      const recordedPrevTag = previousBuildId ? state?.cdnTags?.[previousBuildId] : undefined;
-      if (previousBuildId && recordedPrevTag) {
-        cdnTags[previousBuildId] = recordedPrevTag;
-      }
-      cdnTags[buildId] = cdnTagForBuildId(buildId);
-      // Same shape and pruning as cdnTags: keep the two builds still in play, so a rollback
-      // to either can pin the routing image by digest instead of reconstructing a tag.
-      const routingImageDigests: Record<string, string> = {};
-      const recordedPrevDigest = previousBuildId
-        ? state?.routingImageDigests?.[previousBuildId]
-        : undefined;
-      if (previousBuildId && recordedPrevDigest) {
-        routingImageDigests[previousBuildId] = recordedPrevDigest;
-      }
-      if (imageDigests.routingService) {
-        routingImageDigests[buildId] = imageDigests.routingService;
-      }
-      // The stable routing Deployment is updated in place, so rollback must move its
-      // architecture selector with the target image. Record this build's exact platform and
-      // carry only the outgoing rollback target when its provenance is known.
-      const targetPlatforms: Record<string, TargetPlatform> = {};
-      const recordedPrevPlatform = previousBuildId
-        ? state?.targetPlatforms?.[previousBuildId]
-        : undefined;
-      if (previousBuildId && recordedPrevPlatform) {
-        targetPlatforms[previousBuildId] = recordedPrevPlatform;
-      }
-      targetPlatforms[buildId] = builtTargetPlatform;
-      // N30: carry (and prune to the builds in play) the record of which builds have NO
-      // retained routing manifest, so doctor can qualify "rollback ready" honestly.
-      const unretainedManifestBuilds = [
-        ...(state?.unretainedManifestBuilds ?? []).filter((b) => b === previousBuildId),
-        ...(unretainedManifestBuild ? [unretainedManifestBuild] : []),
-      ].filter((b, i, all) => all.indexOf(b) === i);
-      const compositionPlans: NonNullable<AdapterState["compositionPlans"]> = {};
-      const previousCompositionPlan = previousBuildId
-        ? state?.compositionPlans?.[previousBuildId]
-        : undefined;
-      if (previousBuildId && previousCompositionPlan) {
-        compositionPlans[previousBuildId] = previousCompositionPlan;
-      }
-      if (compositionSnapshot) {
-        compositionPlans[buildId] = {
-          digest: compositionSnapshot.digest,
-          targetFingerprint: compositionSnapshot.plan.target.fingerprint,
-        };
-      }
-      // A previous build with no surviving Deployments (previousPools === []) is recorded
-      // as ABSENT, not as an empty topology: recordedBuildPools rejects [] as malformed on
-      // the next read, and there is nothing for rollback to target anyway.
-      const recordablePreviousBuildId = previousPools.length > 0 ? previousBuildId : undefined;
-      await writeState(
-        projectDir,
+      await runCutover(
         {
+          projectDir,
+          releaseName,
+          namespace,
           buildId,
           previousBuildId,
-          cdnTags,
-          ...(Object.keys(routingImageDigests).length > 0 ? { routingImageDigests } : {}),
-          poolTopologies: {
-            ...(recordablePreviousBuildId
-              ? { [recordablePreviousBuildId]: [...previousPools] }
-              : {}),
-            [buildId]: [...pools],
-          },
-          ...(hasPortableOrigin || state?.defaultPools
-            ? {
-                defaultPools: {
-                  ...(recordablePreviousBuildId
-                    ? {
-                        [recordablePreviousBuildId]:
-                          state?.defaultPools?.[recordablePreviousBuildId] ?? previousPools[0]!,
-                      }
-                    : {}),
-                  [buildId]: defaultPool,
-                },
-              }
-            : {}),
-          targetPlatforms,
-          ...(Object.keys(compositionPlans).length > 0 ? { compositionPlans } : {}),
-          // The build this deploy just installed serves /readyz, so the NEXT deploy can flip
-          // the load balancer's HealthCheckPolicy to readiness without stranding it.
-          readinessPathSupported: true,
-          ...(unretainedManifestBuilds.length > 0 ? { unretainedManifestBuilds } : {}),
-          // N21/N69: the generation this write is based on — writeState treats it as a floor
-          // so the stamped value stays above the cluster's even if the pre-write cluster read
-          // fails (in CI there is no local state.json to carry it forward). Required by the
-          // StateWrite type: rollback used to omit the same value silently.
-          basedOnGeneration: state?.generation ?? null,
-        },
-        releaseName,
-        namespace,
-      );
-    } catch (err) {
-      // N67: the temporary warm-up bounds are scoped to the warm-up on this path too — traffic HAS
-      // switched, so the chart's own bounds are the right ones to autoscale under, and a
-      // deploy that exits here must not leave a raised replica floor for the operator to find.
-      await restoreWarmedHpas();
-      console.error(`\n  Cutover succeeded, but persisting deploy state failed:`);
-      // L14: the error wraps cluster-sourced write/read failures.
-      console.error(`  ${sanitizeForTerminal(err instanceof Error ? err.message : String(err))}`);
-      console.error(
-        `  The new build IS serving, but the cluster ConfigMap was not updated. Restore`,
-      );
-      console.error(
-        `  connectivity and re-run so cluster/local state agree before the next deploy.\n`,
-      );
-      process.exit(1);
-    }
-
-    // 7d-bis. N67: the cutover is durable and the new build is taking 100% of traffic at
-    // the outgoing build's capacity — the warm-up is over, so hand the pool back to the
-    // chart's autoscaling bounds. Deliberately AFTER the state commit (a failure here must
-    // never be able to lose a confirmed cutover) and BEFORE the best-effort steps below, so
-    // the HPA is chart-intended on every path out of this function. Scaling DOWN from here
-    // is the autoscaler's decision under real load, bounded by the operator's config.
-    await restoreWarmedHpas();
-
-    // 7e. State is durable. Now invalidate the PREVIOUS build's Cloud CDN entries
-    // (best-effort, non-fatal; TTL self-heals) so its stale content stops serving.
-    // Deliberately post-cutover — the new origin must be live before the old entries
-    // are dropped — but after the state commit (7d) so a failure here can't leave
-    // cluster state pointing at the outgoing build.
-    const cdnFilterPath = path.join(outputDir, "chart", "templates", "cdn-http-filter.yaml");
-    if (!compositionSnapshot && existsSync(cdnFilterPath) && infra.projectId) {
-      try {
-        await invalidateCdnBuildTag({
+          pools,
+          previousPools,
+          defaultPool,
+          hasPortableOrigin,
+          previousReplicasByPool,
+          state,
+          compositionSnapshot,
+          imageDigests,
+          builtTargetPlatform,
+          unretainedManifestBuild,
           projectId: infra.projectId,
-          releaseName,
           outputDir,
-          buildId: previousBuildId ?? undefined,
-          // M13: the tag recorded when the OUTGOING build deployed — never re-derived
-          // here. Absent (pre-recording state) → cdn-invalidate purges --path=/*.
-          recordedTag: previousBuildId ? state?.cdnTags?.[previousBuildId] : undefined,
-          run: execCapture,
-          log: (m) => console.log(m),
-        });
-      } catch (err) {
-        console.log(
-          `  ! CDN invalidation error (non-fatal): ${err instanceof Error ? err.message : String(err)}`,
-        );
-      }
-    }
-
-    // 7f. State is durable and traffic has switched, so it is now safe to scale the previous
-    // build down to 0. It served through the rollout and remains as the rollback target.
-    // Best-effort: success is already durable (7d) — a failure here warns instead of
-    // failing the whole deploy (the previous build just keeps burning replicas until the
-    // next deploy or a manual scale-down). The outgoing HPA was transferred out of Helm
-    // release-manifest lifecycle before the upgrade so it could remain active while this build
-    // served. Delete it now and wait for that deletion to finish BEFORE setting replicas=0;
-    // if deletion fails, leave the Deployment alone rather than asking a still-live HPA to
-    // fight the scale command.
-    if (previousBuildId && previousBuildId !== buildId) {
-      let scaleDownFailed = false;
-      for (const poolName of previousPools) {
-        // The HPA name must come from the SAME suffix-reserving helper the template
-        // uses (hpa.ts truncates the base at 59, then appends "-hpa") — concatenating
-        // "-hpa" onto the 63-truncated deployment name diverges past that boundary
-        // (and can exceed 63 chars), so this delete silently missed the real HPA and
-        // the autoscaler rescaled the parked previous build right back up.
-        const { deployment: poolPrevName, hpa: poolPrevHpa } = poolResourceNames(
-          releaseName,
-          poolName,
-          previousBuildId,
-        );
-        const hpaDelete = await execCapture(
-          "kubectl",
-          ["delete", "hpa", poolPrevHpa, "-n", namespace, "--ignore-not-found"],
-          { timeoutMs: EXEC_TIMEOUTS.kubectl },
-        );
-        if (hpaDelete.exitCode !== 0) {
-          scaleDownFailed = true;
-          console.warn(
-            `  ! Could not remove outgoing autoscaler ${poolPrevHpa}; leaving ` +
-              `${poolPrevName} at its current replica count because that HPA could immediately ` +
-              `undo a scale to zero: ${sanitizeForTerminal(hpaDelete.stderr.trim()) || "unknown error"}`,
-          );
-          continue;
-        }
-        const scaleDown = await execCapture(
-          "kubectl",
-          ["scale", `deployment/${poolPrevName}`, "-n", namespace, "--replicas=0"],
-          { timeoutMs: EXEC_TIMEOUTS.kubectl },
-        );
-        if (scaleDown.exitCode !== 0) {
-          scaleDownFailed = true;
-          console.warn(
-            `  ! Could not scale down ${poolPrevName}: ` +
-              `${sanitizeForTerminal(scaleDown.stderr.trim()) || "unknown error"}`,
-          );
-        }
-      }
-      if (scaleDownFailed) {
-        console.warn(
-          `  ! Previous build ${previousBuildId} was NOT fully scaled to 0 — it keeps its ` +
-            `replicas until the next deploy. Scale it down manually when convenient.`,
-        );
-      } else {
-        console.log(`  → Previous build scaled to 0 (kept for rollback)`);
-      }
-    }
-
-    // 7g. Clean up old deployments.
-    // The previous build was scaled to 0 in step 7f above (kept as the rollback target).
-    // Delete anything that isn't the current or previous build. Classify by EXACT deployment name
-    // (reconstructed with the same sanitizer the template uses) rather than a 12-char normalized
-    // substring — a shared prefix between two build ids could otherwise delete the wrong build.
-    const currentDeployNames = new Set(
-      pools.map((p) => sanitizeK8sName(`${releaseName}-${p}-${buildId}`)),
-    );
-    const previousDeployNames = previousBuildId
-      ? new Set(previousPools.map((p) => sanitizeK8sName(`${releaseName}-${p}-${previousBuildId}`)))
-      : undefined;
-
-    const allDeploys = await execCapture(
-      "kubectl",
-      [
-        "get",
-        "deployments",
-        "-n",
-        namespace,
-        "-l",
-        `app.kubernetes.io/name=${releaseName}`,
-        "-o",
-        'jsonpath={range .items[*]}{.metadata.name}{"\\n"}{end}',
-      ],
-      { timeoutMs: EXEC_TIMEOUTS.kubectl },
-    );
-    if (allDeploys.exitCode === 0) {
-      for (const name of allDeploys.stdout.trim().split("\n")) {
-        // The routing tier is a stable in-place Deployment — never a cleanup candidate.
-        // EXACT name match: a substring match would also spare (or, elsewhere, drop) a
-        // pool deployment that merely contains "routing-service" in its name.
-        if (!name || name === routingDeploy) continue;
-        // Keep current build
-        if (currentDeployNames.has(name)) continue;
-        // Keep previous build (rollback target)
-        if (previousDeployNames?.has(name)) {
-          console.log(`  → Previous build kept for rollback: ${name}`);
-          continue;
-        }
-        // If we don't know the previous build, we cannot classify this non-current
-        // deployment as safe to delete. Be conservative and keep it as a potential
-        // rollback target rather than blanket-deleting every non-current build.
-        if (!previousBuildId) {
-          console.log(
-            `  → Conservative cleanup: keeping non-current build "${name}" ` +
-              `(previous-build state unknown, cannot classify as safe to delete)`,
-          );
-          continue;
-        }
-        // Delete everything else
-        console.log(`  → Deleting old build: ${name}`);
-        await execCapture("kubectl", ["delete", "deployment", name, "-n", namespace], {
-          timeoutMs: EXEC_TIMEOUTS.kubectl,
-        });
-        await execCapture("kubectl", ["delete", "service", name, "-n", namespace], {
-          timeoutMs: EXEC_TIMEOUTS.kubectl,
-        }).catch(() => {});
-        // This build's raw release/pool/buildId parts are unknowable here (`name` is a
-        // cluster-listed deployment of a build state no longer tracks), so the HCP name
-        // can't go through poolResourceNames. Re-sanitizing the deployment name with the
-        // "-hcp" suffix reproduces the template's name exactly: for a base past the
-        // 59-char boundary, re-truncating the 63-char deployment name to 59 yields the
-        // same prefix the template truncated to (any trailing hyphens the template's
-        // strip removed beyond 59 are re-stripped here). Bare `${name}-hcp` diverged
-        // there — and could exceed 63 chars, an invalid name kubectl rejects.
-        await execCapture(
-          "kubectl",
-          ["delete", "healthcheckpolicy", sanitizeK8sName(name, "-hcp"), "-n", namespace],
-          { timeoutMs: EXEC_TIMEOUTS.kubectl },
-        ).catch(() => {});
-      }
-    }
-
-    // Stable resources transferred out of Helm's deletion set deliberately survive without
-    // rewriting their managed-by label while their pool belongs to either build in play. Once a
-    // later successful cutover removes that pool from BOTH topologies, delete the exact
-    // adapter-retained Service/PDB/HCP group. This is post-state
-    // commit and best-effort: cleanup failure leaks bounded objects but cannot invalidate the
-    // serving cutover. The helper validates ownership, labels, names, and selectors before any
-    // deletion and keeps the Service as a retry anchor when a companion delete fails.
-    try {
-      let healthCheckPolicyCrd = false;
-      try {
-        healthCheckPolicyCrd = await hasHealthCheckPolicyCrd();
-      } catch (err) {
-        // Post-commit cleanup can still safely classify Service/PDB without cluster-wide CRD
-        // permission. Preserve a truthful warning that HCP cleanup was skipped.
-        console.warn(
-          `  ! Could not classify retained HealthCheckPolicy objects: ` +
-            `${err instanceof Error ? err.message : String(err)}. Service/PDB cleanup will ` +
-            `continue; any retained HCP was left in place.`,
-        );
-      }
-      const stableCleanup = await cleanupRetainedStablePoolResources({
-        releaseName,
-        namespace,
-        keepPools: new Set([...pools, ...previousPools]),
-        healthCheckPolicyCrd,
-      });
-      for (const deleted of stableCleanup.deleted) {
-        console.log(`  → Deleted obsolete retained stable resource: ${deleted}`);
-      }
-      for (const failure of stableCleanup.failures) {
-        console.warn(`  ! Retained stable-resource cleanup incomplete: ${failure}`);
-      }
-    } catch (err) {
-      console.warn(
-        `  ! Retained stable-resource cleanup could not run: ` +
-          `${err instanceof Error ? err.message : String(err)}. The deploy is committed; ` +
-          `the retained objects were left in place for a later retry.`,
-      );
-    }
-
-    // Clean up OLD route-ext Jobs (K8s Jobs are immutable; each deploy creates a fresh
-    // one). Skip the CURRENT job by EXACT name — a fuzzy build-id substring match here
-    // (12-char slice vs the name's 10-char slice) previously deleted the running job
-    // before it could register the extension, so the traffic ext never got reconciled.
-    const oldJobs = await execCapture(
-      "kubectl",
-      [
-        "get",
-        "jobs",
-        "-n",
-        namespace,
-        "-l",
-        `app.kubernetes.io/name=${releaseName},app.kubernetes.io/component=route-ext-job`,
-        "-o",
-        'jsonpath={range .items[*]}{.metadata.name}{"\\n"}{end}',
-      ],
-      { timeoutMs: EXEC_TIMEOUTS.kubectl },
-    );
-    if (oldJobs.exitCode === 0) {
-      for (const jobName of oldJobs.stdout.trim().split("\n")) {
-        if (!jobName || jobName === currentRouteExtJob) continue;
-        await execCapture("kubectl", ["delete", "job", jobName, "-n", namespace], {
-          timeoutMs: EXEC_TIMEOUTS.kubectl,
-        });
-      }
-    }
-
-    // Prune retained routing-manifest snapshot ConfigMaps that belong to neither the
-    // current nor the previous build. Deploy and rollback each retain one (up to ~1 MiB
-    // of etcd apiece) and nothing else ever deleted them, so they accumulated
-    // unboundedly. Classify by EXACT snapshot name (the same helper that named them);
-    // mirror the deployment cleanup's conservatism above — with no known previous
-    // build, non-current snapshots can't be classified as safe to delete, so keep them.
-    // Best-effort: state is already committed (7d), a failure here only leaks storage.
-    if (previousBuildId) {
-      const keepSnapshots = new Set([
-        routingManifestSnapshotName(releaseName, buildId),
-        routingManifestSnapshotName(releaseName, previousBuildId),
-      ]);
-      const snapshots = await execCapture(
-        "kubectl",
-        [
-          "get",
-          "configmaps",
-          "-n",
-          namespace,
-          "-l",
-          `app.kubernetes.io/name=${releaseName},` +
-            `app.kubernetes.io/component=${ROUTING_MANIFEST_SNAPSHOT_COMPONENT}`,
-          "-o",
-          'jsonpath={range .items[*]}{.metadata.name}{"\\n"}{end}',
-        ],
-        { timeoutMs: EXEC_TIMEOUTS.kubectl },
-      );
-      if (snapshots.exitCode === 0) {
-        for (const cmName of snapshots.stdout.trim().split("\n")) {
-          if (!cmName || keepSnapshots.has(cmName)) continue;
-          console.log(`  → Deleting old routing-manifest snapshot: ${cmName}`);
-          await execCapture("kubectl", ["delete", "configmap", cmName, "-n", namespace], {
-            timeoutMs: EXEC_TIMEOUTS.kubectl,
-          });
-        }
-      }
-
-      if (hasPortableOrigin) {
-        const keepPlans = new Set([
-          compositionPlanConfigMapName(releaseName, buildId),
-          compositionPlanConfigMapName(releaseName, previousBuildId),
-        ]);
-        const plans = await execCapture(
-          "kubectl",
-          [
-            "get",
-            "configmaps",
-            "-n",
-            namespace,
-            "-l",
-            `app.kubernetes.io/name=${releaseName},app.kubernetes.io/component=${COMPOSITION_PLAN_COMPONENT}`,
-            "-o",
-            'jsonpath={range .items[*]}{.metadata.name}{"\\n"}{end}',
-          ],
-          { timeoutMs: EXEC_TIMEOUTS.kubectl },
-        );
-        if (plans.exitCode === 0) {
-          for (const cmName of plans.stdout.trim().split("\n")) {
-            if (!cmName || keepPlans.has(cmName)) continue;
-            console.log(`  → Deleting old composition plan: ${cmName}`);
-            await execCapture("kubectl", ["delete", "configmap", cmName, "-n", namespace], {
-              timeoutMs: EXEC_TIMEOUTS.kubectl,
-            });
-          }
-        }
-      }
-    }
-
-    // N87. Prune internal dispatch Secrets that no Deployment references any more. Each
-    // build renders its OWN Secret (`<release>-ihs-<build>-<digest>`) annotated
-    // `helm.sh/resource-policy: keep`, so helm never prunes them — deliberately: the
-    // retained previous build's pods reference theirs BY NAME and a missing secretKeyRef is
-    // CreateContainerConfigError, so pruning on upgrade would brick both a restart in the
-    // deploy window and the rollback. Keeping them makes GC this step's job (the
-    // routing-manifest snapshots accumulated unboundedly for exactly the same reason).
-    //
-    // Classify by what is actually REFERENCED, not by build id: that covers the current
-    // build, the retained previous build, the LEGACY stable-named Secret still needed by a
-    // pre-N87 rollback target, the builds whose Deployments were just deleted above, and
-    // leftovers from an earlier incarnation of this release name — with no assumption about
-    // which builds exist. Conservative on every failure: an unreadable or unparseable
-    // reference set skips the sweep entirely, because deleting a Secret a live pod template
-    // needs would brick that build's restarts. Best-effort — state is already committed
-    // (7d), so a failure here only leaks a handful of 64-byte Secrets.
-    const deploySpecs = await execCapture(
-      "kubectl",
-      [
-        "get",
-        "deployments",
-        "-n",
-        namespace,
-        "-l",
-        `app.kubernetes.io/name=${releaseName}`,
-        "-o",
-        "json",
-      ],
-      { timeoutMs: EXEC_TIMEOUTS.kubectl },
-    );
-    let referencedSecrets: Set<string> | null = null;
-    if (deploySpecs.exitCode === 0 && deploySpecs.stdout.trim()) {
-      try {
-        const parsed = JSON.parse(deploySpecs.stdout) as {
-          items?: {
-            spec?: {
-              template?: {
-                spec?: {
-                  containers?: { env?: { valueFrom?: { secretKeyRef?: { name?: string } } }[] }[];
-                };
-              };
-            };
-          }[];
-        };
-        // Every secretKeyRef, not only the internal one — a superset is harmless here
-        // because only Secrets carrying the internal-secret component label are candidates.
-        referencedSecrets = new Set(
-          (parsed.items ?? []).flatMap((d) =>
-            (d.spec?.template?.spec?.containers ?? []).flatMap((c) =>
-              (c.env ?? [])
-                .map((e) => e.valueFrom?.secretKeyRef?.name)
-                .filter((n): n is string => typeof n === "string" && n.length > 0),
-            ),
+          hasRouteExtJob: existsSync(
+            path.join(outputDir, "chart", "templates", "route-ext-update-job.yaml"),
           ),
-        );
-        // Belt and braces: this deploy's own Secret is referenced by definition, and the
-        // pod templates it belongs to may not be listed yet on a slow API read.
-        referencedSecrets.add(internalSecretName(releaseName, buildId));
-        if (previousBuildId) {
-          referencedSecrets.add(internalSecretName(releaseName, previousBuildId));
-        }
-      } catch {
-        referencedSecrets = null;
-      }
-    }
-    if (referencedSecrets === null) {
-      console.warn(
-        `  ! Could not read which Secrets the release's Deployments reference ` +
-          `(kubectl exit ${deploySpecs.exitCode}) — skipping the internal-secret prune. ` +
-          `Old per-build dispatch Secrets stay until the next successful deploy.`,
+          hasEnvoyExtensionPolicy: existsSync(
+            path.join(outputDir, "chart", "templates", "envoy-extension-policy.yaml"),
+          ),
+          cdnEnabled: chartHasCdn,
+        },
+        { restoreEdgeToPreviousBuild, edgeStatusLines },
       );
-    } else {
-      const internalSecrets = await execCapture(
-        "kubectl",
-        [
-          "get",
-          "secrets",
-          "-n",
-          namespace,
-          "-l",
-          `app.kubernetes.io/name=${releaseName},` +
-            `app.kubernetes.io/component=${INTERNAL_SECRET_COMPONENT}`,
-          "-o",
-          'jsonpath={range .items[*]}{.metadata.name}{"\\n"}{end}',
-        ],
-        { timeoutMs: EXEC_TIMEOUTS.kubectl },
-      );
-      if (internalSecrets.exitCode === 0) {
-        for (const secretName of internalSecrets.stdout.trim().split("\n")) {
-          if (!secretName || referencedSecrets.has(secretName)) continue;
-          console.log(`  → Deleting unreferenced internal dispatch Secret: ${secretName}`);
-          await execCapture("kubectl", ["delete", "secret", secretName, "-n", namespace], {
-            timeoutMs: EXEC_TIMEOUTS.kubectl,
-          });
-        }
-      }
+    } catch (err) {
+      if (err instanceof CutoverExitError) process.exit(err.code);
+      throw err;
     }
   }
